requirements.txt
-e .

pytest-django
factory_boy
django-stubs
tox==3.24.4
build==0.7.0
